Rename directory: ruby_scripts/ -> scripts/
